+++ lib/read/cardano-wallet-read.cabal
@@ -47,6 +47,7 @@ library
     Cardano.Read.Ledger.Block.Block
     Cardano.Read.Ledger.Block.BHeader
     Cardano.Read.Ledger.Block.BlockNo
+    Cardano.Read.Ledger.Block.HeaderHash
     Cardano.Read.Ledger.Tx.Inputs
     Cardano.Read.Ledger.Tx.Outputs
     Cardano.Read.Ledger.Tx.TxId
@@ -76,7 +77,6 @@ library
     Cardano.Wallet.Read.Block.Gen.Build
     Cardano.Wallet.Read.Block.Gen.Byron
     Cardano.Wallet.Read.Block.Gen.Shelley
-    Cardano.Wallet.Read.Block.HeaderHash
     Cardano.Wallet.Read.Block.SlotNo
     Cardano.Wallet.Read.Block.Txs
     Cardano.Wallet.Read.Chain
